default cont へテスト用の関数追加、SQLに関数の外枠追加
